feat: check data sensitivity in excel workbook
