+++ packages/backend/src/apps/m365-excel/common/sharepoint/excel-workbook/constants.ts
@@ -17,3 +17,6 @@ export interface ExcelTableColumn extends IJSONObject {
 
 // 30 min; see ExcelWorkbook's comment for context.
 export const DEFAULT_CACHE_LIFETIME_SECONDS = 30 * 60
+
+// 30 min; see ExcelWorkbook's comment for context.
+export const SENSITIVITY_LABEL_CACHE_LIFETIME_SECONDS = 60
